@@ -61,63 +61,57 @@ public class SecurityController : Controller
 using Microsoft.AspNetCore.Mvc;
 using SS.Backend.Security;
 
-namespace AuthAPI.Controllers
+namespace Security_Controller.Controllers;
+
+[ApiController]
+[Route("[controller]")]
+public class SecurityController : Controller
 {
-    [ApiController]
-    [Route("api/auth")]
-    public class AuthNController : Controller
+    private readonly SSAuthService _authService;
+    public SecurityController(SSAuthService authService)
     {
-        private readonly SSAuthService _authService;
+        _authService = authService;
+    }
 
-        public AuthNController(SSAuthService authService)
+    [HttpPost("sendOTP")]
+    public async Task<IActionResult> SendOTP([FromBody] AuthenticationRequest request)
+    {
+        var (otp, response) = await _authService.SendOTP_and_SaveToDB(request);
+
+        if (response.HasError)
         {
-            _authService = authService;
+            return BadRequest(response.ErrorMessage);
         }
 
-        [HttpPost("sendOTP")]
-        public async Task<IActionResult> SendOTP([FromBody] AuthenticationRequest request)
+        return Ok(new { otp });
+    }
+
+    [HttpPost("authenticate")]
+    public async Task<IActionResult> Authenticate([FromBody] AuthenticationRequest request)
+    {
+        var (principal, response) = await _authService.Authenticate(request);
+
+        if (response.HasError)
         {
-            var (otp, response) = await _authService.SendOTP_and_SaveToDB(request);
-
-            if (response.HasError)
-            {
-                return BadRequest(response.ErrorMessage);
-            }
-
-            return Ok(new { otp });
+            return BadRequest(response.ErrorMessage);
         }
 
-        [HttpPost("authenticate")]
-        public async Task<IActionResult> Authenticate([FromBody] AuthenticationRequest request)
+        return Ok(new { principal });
+    }
+
+    [HttpPost(Name = "postAuthZ")]
+    [Route("/authZ")]
+    public async Task<IActionResult> AuthZ([FromQueryAttribute] SSPrincipal currentPrincipal, IDictionary<string, string> requiredClaims)
+    {
+        bool verify = await _authService.IsAuthorize(currentPrincipal, requiredClaims);
+
+        if (!verify)
         {
-            var (principal, response) = await _authService.Authenticate(request);
-
-            if (response.HasError)
-            {
-                return BadRequest(response.ErrorMessage);
-            }
-
-            return Ok(new { principal });
+            // If not authorized, return 401 Unauthorized
+            return Unauthorized();
         }
-
-        [HttpGet(Name = "getAuthZ")]
-        [Route("/authZ")]
-        public async Task<IActionResult> AuthZ()
-        {
-            var currentPrincipal = HttpContext.User;
-            IDictionary<string, string> requiredClaims = new Dictionary<string, string>
-            {
-                {"Role", "Admin"}
-            }
-            var verify = await _authService.IsAuthorize(currentPrincipal, requiredClaims);
-
-             if (!verify)
-            {
-                // If not authorized, return 401 Unauthorized
-                return Unauthorized();
-            }
-            return Ok(new { verify} );
-        }
+        return Ok(new { verify} );
     }
 }
+
 
