@@ -33,9 +33,9 @@ builder.Services.AddAuthentication(options =>
 
 builder.Services.AddAuthorization();
 
-//builder.Services.AddTransient<ConfigService>(provider =>
-  //  new ConfigService(Path.Combine(AppContext.BaseDirectory, "config.local.txt")));
-//builder.Services.AddTransient<SqlDAO>();
+builder.Services.AddTransient<ConfigService>(provider =>
+    new ConfigService(Path.Combine("/Users/sarahsantos/SpaceSurfer/Configs/config.local.txt")));//AppContext.BaseDirectory, "config.local.txt")));
+builder.Services.AddTransient<SqlDAO>();
 builder.Services.AddTransient<ISqlDAO, SqlDAO>();
 builder.Services.AddTransient<CustomSqlCommandBuilder>();
 builder.Services.AddTransient<GenOTP>();
